Rename SDK package name from com.mozilla.mentat to org.mozilla.mentat
