--- sdks/android/Mentat/library/src/main/java/com/mozilla/mentat/EntityBuilder.java
@@ -1,357 +1,0 @@
-/* -*- Mode: Java; c-basic-offset: 4; tab-width: 20; indent-tabs-mode: nil; -*-
- * Copyright 2018 Mozilla
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use
- * this file except in compliance with the License. You may obtain a copy of the
- * License at http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software distributed
- * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
- * CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License. */
-
-package com.mozilla.mentat;
-
-import android.util.Log;
-
-import com.sun.jna.Memory;
-import com.sun.jna.Pointer;
-
-import java.io.IOException;
-import java.nio.ByteBuffer;
-import java.util.Date;
-import java.util.UUID;
-
-/**
- * This class wraps a raw pointer that points to a Rust `EntityBuilder<InProgressBuilder>` object.
- *
- * {@link EntityBuilder} provides a programmatic interface to performing assertions on a specific entity.
- * It provides functions for adding and retracting values for attributes for an entity within
- * an in progress transaction.
- * <p/>
- * The `transact` function will transact the assertions that have been added to the {@link EntityBuilder}
- * and pass back the {@link TxReport} that was generated by this transact and the {@link InProgress} that was
- * used to perform the transact. This enables you to perform further transacts on the same {@link InProgress}
- * before committing.
- * <p/>
- * <pre>{@code
- * long aEntid = txReport.getEntidForTempId("a");
- * long bEntid = txReport.getEntidForTempId("b");
- * EntityBuilder builder = mentat.getEntityBuilderForEntid(bEntid);
- * builder.add(":foo/boolean", true);
- * builder.add(":foo/instant", newDate);
- * InProgress inProgress = builder.transact().getInProgress();
- * inProgress.transact("[[:db/add \(aEntid) :foo/long 22]]");
- * inProgress.commit();
- * }</pre>
- * <p/>
- * The `commit` function will transact and commit the assertions that have been added to the {@link EntityBuilder}.
- * It will consume the {@link InProgress} used to perform the transact. It returns the {@link TxReport} generated by
- * the transact. After calling `commit`, a new transaction must be started by calling <pre>Mentat.beginTransaction()</pre>
- * in order to perform further actions.
- * <p/>
- * <pre>{@code
- * long aEntid = txReport.getEntidForTempId("a");
- * EntityBuilder builder = mentat.getEntityBuilderForEntid(bEntid);
- * builder.add(":foo/boolean", true);
- * builder.add(":foo/instant", newDate);
- * builder.commit();
- * }</pre>
- */
-public class EntityBuilder extends RustObject {
-
-    public EntityBuilder(Pointer pointer) {
-        this.rawPointer = pointer;
-    }
-
-    /**
-     * Asserts the value of attribute `keyword` to be the provided `value`.
-     * @param keyword   The name of the attribute in the format `:namespace/name`.
-     * @param value The value to be asserted
-     */
-    public void add(String keyword, long value) {
-        this.validate();
-        RustResult result = JNA.INSTANCE.entity_builder_add_long(this.rawPointer, keyword, value);
-        if (result.isFailure()) {
-            Log.e("EntityBuilder", result.err);
-        }
-    }
-
-    /**
-     * Asserts the value of attribute `keyword` to be the provided `value`.
-     * // TODO throw exception if error occurs
-     * @param keyword   The name of the attribute in the format `:namespace/name`.
-     * @param value The value to be asserted
-     */
-    public void addRef(String keyword, long value) {
-        this.validate();
-        RustResult result = JNA.INSTANCE.entity_builder_add_ref(this.rawPointer, keyword, value);
-        if (result.isFailure()) {
-            Log.e("EntityBuilder", result.err);
-        }
-    }
-
-    /**
-     * Asserts the value of attribute `keyword` to be the provided `value`.
-     * // TODO throw exception if error occurs
-     * @param keyword   The name of the attribute in the format `:namespace/name`.
-     * @param value The value to be asserted
-     */
-    public void addKeyword(String keyword, String value) {
-        this.validate();
-        RustResult result = JNA.INSTANCE.entity_builder_add_keyword(this.rawPointer, keyword, value);
-        if (result.isFailure()) {
-            Log.e("EntityBuilder", result.err);
-        }
-    }
-
-    /**
-     * Asserts the value of attribute `keyword` to be the provided `value`.
-     * // TODO throw exception if error occurs
-     * @param keyword   The name of the attribute in the format `:namespace/name`.
-     * @param value The value to be asserted
-     */
-    public void add(String keyword, boolean value) {
-        this.validate();
-        RustResult result = JNA.INSTANCE.entity_builder_add_boolean(this.rawPointer, keyword, value ? 1 : 0);
-        if (result.isFailure()) {
-            Log.e("EntityBuilder", result.err);
-        }
-    }
-
-    /**
-     * Asserts the value of attribute `keyword` to be the provided `value`.
-     * // TODO throw exception if error occurs
-     * @param keyword   The name of the attribute in the format `:namespace/name`.
-     * @param value The value to be asserted
-     */
-    public void add(String keyword, double value) {
-        this.validate();
-        RustResult result = JNA.INSTANCE.entity_builder_add_double(this.rawPointer, keyword, value);
-        if (result.isFailure()) {
-            Log.e("EntityBuilder", result.err);
-        }
-    }
-
-    /**
-     * Asserts the value of attribute `keyword` to be the provided `value`.
-     * // TODO throw exception if error occurs
-     * @param keyword   The name of the attribute in the format `:namespace/name`.
-     * @param value The value to be asserted
-     */
-    public void add(String keyword, Date value) {
-        this.validate();
-        RustResult result = JNA.INSTANCE.entity_builder_add_timestamp(this.rawPointer, keyword, value.getTime() * 1_000);
-        if (result.isFailure()) {
-            Log.e("EntityBuilder", result.err);
-        }
-    }
-
-    /**
-     * Asserts the value of attribute `keyword` to be the provided `value`.
-     * // TODO throw exception if error occurs
-     * @param keyword   The name of the attribute in the format `:namespace/name`.
-     * @param value The value to be asserted
-     */
-    public void add(String keyword, String value) {
-        this.validate();
-        RustResult result = JNA.INSTANCE.entity_builder_add_string(this.rawPointer, keyword, value);
-        if (result.isFailure()) {
-            Log.e("EntityBuilder", result.err);
-        }
-    }
-
-    /**
-     * Asserts the value of attribute `keyword` to be the provided `value`.
-     * // TODO throw exception if error occurs
-     * @param keyword   The name of the attribute in the format `:namespace/name`.
-     * @param value The value to be asserted
-     */
-    public void add(String keyword, UUID value) {
-        this.validate();
-
-        RustResult result = JNA.INSTANCE.entity_builder_add_uuid(this.rawPointer, keyword, getPointerForUUID(value));
-        if (result.isFailure()) {
-            Log.e("EntityBuilder", result.err);
-        }
-    }
-
-    /**
-     * Retracts the value of attribute `keyword` from the provided `value`.
-     *
-     * TODO throw exception if error occurs
-     *
-     * @param keyword   The name of the attribute in the format `:namespace/name`.
-     * @param value The value to be retracted
-     */
-    public void retract(String keyword, long value) {
-        this.validate();
-        RustResult result = JNA.INSTANCE.entity_builder_retract_long(this.rawPointer, keyword, value);
-        if (result.isFailure()) {
-            Log.e("EntityBuilder", result.err);
-        }
-    }
-
-
-    /**
-     * Retracts the value of attribute `keyword` from the provided `value`.
-     *
-     * TODO throw exception if error occurs
-     *
-     * @param keyword   The name of the attribute in the format `:namespace/name`.
-     * @param value The value to be retracted
-     */
-    public void retractRef(String keyword, long value) {
-        this.validate();
-        RustResult result = JNA.INSTANCE.entity_builder_retract_ref(this.rawPointer, keyword, value);
-        if (result.isFailure()) {
-            Log.e("EntityBuilder", result.err);
-        }
-    }
-
-    /**
-     * Retracts the value of attribute `keyword` from the provided `value`.
-     *
-     * TODO throw exception if error occurs
-     *
-     * @param keyword   The name of the attribute in the format `:namespace/name`.
-     * @param value The value to be retracted
-     */
-    public void retractKeyword(String keyword, String value) {
-        this.validate();
-        RustResult result = JNA.INSTANCE.entity_builder_retract_keyword(this.rawPointer, keyword, value);
-        if (result.isFailure()) {
-            Log.e("EntityBuilder", result.err);
-        }
-    }
-
-    /**
-     * Retracts the value of attribute `keyword` from the provided `value`.
-     *
-     * TODO throw exception if error occurs
-     *
-     * @param keyword   The name of the attribute in the format `:namespace/name`.
-     * @param value The value to be retracted
-     */
-    public void retract(String keyword, boolean value) {
-        this.validate();
-        RustResult result = JNA.INSTANCE.entity_builder_retract_boolean(this.rawPointer, keyword, value ? 1 : 0);
-        if (result.isFailure()) {
-            Log.e("EntityBuilder", result.err);
-        }
-    }
-
-    /**
-     * Retracts the value of attribute `keyword` from the provided `value`.
-     *
-     * TODO throw exception if error occurs
-     *
-     * @param keyword   The name of the attribute in the format `:namespace/name`.
-     * @param value The value to be retracted
-     */
-    public void retract(String keyword, double value) {
-        this.validate();
-        RustResult result = JNA.INSTANCE.entity_builder_retract_double(this.rawPointer, keyword, value);
-        if (result.isFailure()) {
-            Log.e("EntityBuilder", result.err);
-        }
-    }
-
-    /**
-     * Retracts the value of attribute `keyword` from the provided `value`.
-     *
-     * TODO throw exception if error occurs
-     *
-     * @param keyword   The name of the attribute in the format `:namespace/name`.
-     * @param value The value to be retracted
-     */
-    public void retract(String keyword, Date value) {
-        this.validate();
-        RustResult result = JNA.INSTANCE.entity_builder_retract_timestamp(this.rawPointer, keyword, value.getTime() * 1_000);
-        if (result.isFailure()) {
-            Log.e("EntityBuilder", result.err);
-        }
-    }
-
-    /**
-     * Retracts the value of attribute `keyword` from the provided `value`.
-     *
-     * TODO throw exception if error occurs
-     *
-     * @param keyword   The name of the attribute in the format `:namespace/name`.
-     * @param value The value to be retracted
-     */
-    public void retract(String keyword, String value) {
-        this.validate();
-        RustResult result = JNA.INSTANCE.entity_builder_retract_string(this.rawPointer, keyword, value);
-        if (result.isFailure()) {
-            Log.e("EntityBuilder", result.err);
-        }
-    }
-
-    /**
-     * Retracts the value of attribute `keyword` from the provided `value`.
-     *
-     * TODO throw exception if error occurs
-     *
-     * @param keyword   The name of the attribute in the format `:namespace/name`.
-     * @param value The value to be retracted
-     */
-    public void retract(String keyword, UUID value) {
-        this.validate();
-        RustResult result = JNA.INSTANCE.entity_builder_retract_uuid(this.rawPointer, keyword, this.getPointerForUUID(value));
-        if (result.isFailure()) {
-            Log.e("EntityBuilder", result.err);
-        }
-    }
-
-    /**
-     * Transacts the added assertions. This consumes the pointer associated with this {@link EntityBuilder}
-     * such that no further assertions can be added after the `transact` has completed. To perform
-     * further assertions, use the {@link InProgress} inside the {@link InProgressTransactionResult}
-     * returned from this function.
-     * <p/>
-     * This does not commit the transaction. In order to do so, `commit` can be called on the
-     * {@link InProgress} inside the {@link InProgressTransactionResult} returned from this function.
-     *
-     * TODO throw exception if error occurs
-     *
-     * @return A {@link InProgressTransactionResult} containing the current {@link InProgress} and
-     * the {@link TxReport}  generated by the transact.
-     */
-    public InProgressTransactionResult transact() {
-        this.validate();
-        InProgressTransactionResult result = JNA.INSTANCE.entity_builder_transact(this.rawPointer);
-        this.rawPointer = null;
-        return result;
-    }
-
-    /**
-     * Transacts the added assertions and commits. This consumes the pointer associated with this
-     * {@link EntityBuilder} and the associated {@link InProgress} such that no further assertions
-     * can be added after the `commit` has completed.
-     * <p/>
-     * To perform further assertions, a new `{@link InProgress} or {@link EntityBuilder} should be
-     * created.
-     *
-     * TODO throw exception if error occurs
-     *
-     * @return
-     */
-    public TxReport commit() {
-        this.validate();
-        RustResult result = JNA.INSTANCE.entity_builder_commit(this.rawPointer);
-        this.rawPointer = null;
-        if (result.isFailure()) {
-            Log.e("EntityBuilder", result.err);
-            return null;
-        }
-
-        return new TxReport(result.ok);
-    }
-
-    @Override
-    public void close() throws IOException {
-        if (this.rawPointer != null) {
-            JNA.INSTANCE.entity_builder_destroy(this.rawPointer);
-        }
-    }
-}
